Make items directory and move item-related content to it \ Move user-related content to user directory \ Update folder paths and improts
Remove unneeded auth/auth service
