+++ public/itemizer-app/src/app/app.module.ts
@@ -4,13 +4,12 @@ import { NgModule } from '@angular/core';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { HeaderComponent } from './header/header.component';
-import { ItemListComponent } from './item-list/item-list.component';
+import { ItemListComponent } from './items/item-list/item-list.component';
 import { HttpClientModule } from '@angular/common/http';
-import { ItemService } from './item.service';
-import { AddItemComponent } from './add-item/add-item.component';
+import { ItemService } from './items/item.service';
+import { AddItemComponent } from './items/add-item/add-item.component';
 import { FormsModule } from '@angular/forms';
-import { UserService } from './user.service';
-import { RouterModule } from '@angular/router';
+import { UserService } from './user/user.service';
 import { UserComponent } from './user/user.component';
 
 @NgModule({
