+++ public/itemizer-app/src/app/header/header.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
-import { UserService } from '../user.service';
+import { UserService } from '../user/user.service';
 import { Router } from '@angular/router';
 import { Subscription } from 'rxjs';
-import { User } from '../user.model';
+import { User } from '../user/user.model';
 
 @Component({
   selector: 'app-header',
