+++ public/itemizer-app/src/app/items/item-list/item-list.component.ts
