+++ public/itemizer-app/src/app/items/item.model.ts
