+++ public/itemizer-app/src/app/user/auth.guard.ts
