+++ public/itemizer-app/src/app/user/user.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
-import { UserService } from '../user.service';
+import { UserService } from './user.service';
 import { NgForm } from '@angular/forms';
 import { Router } from '@angular/router';
-import { User } from '../user.model';
+import { User } from './user.model';
 
 @Component({
   selector: 'app-user',
