+++ public/itemizer-app/src/app/user/user.model.ts
